now it's kinda working

+++ db/migrate/20211207072337_create_purchases.rb
@@ -4,8 +4,8 @@ class CreatePurchases < ActiveRecord::Migration[5.2]
       t.integer :supplier_id
       t.date :date_of_the_order
       t.date :expected_date_of_delivery
-      t.float :estimated_total
-      t.float :actual_total
+      t.decimal :estimated_total
+      t.decimal :actual_total
 
       t.timestamps
     end
